[develop]: Add Tutorial Chapter for the SRW App (#556)
This PR adds a Tutorial chapter to the SRW App (and several plotting/png files, which makes this PR look much bigger than it is!). The Tutorial chapter outlines 5 severe weather cases. A complete tutorial is available for the first case. The hope is to fill out the others with each new SRW App release. To supplement the tutorial, a document on SSH access and data transfer was added to help users download plots from an HPC/cloud system to their local device. Additionally, a few Glossary updates were needed for the tutorial, as well as minor corrections to the plotting/TemplateVars explanations.

---------

Co-authored-by: gspetro <gillian.s.petro@gmail.com>